refactor(): BridgeRepository -> Bridge 이름 수정및 repository 패키지에서 domain 패키지로 이동
