+++ src/main/java/bridge/config/AppConfig.java
@@ -2,9 +2,9 @@ package bridge.config;
 
 import bridge.controller.BridgeGame;
 import bridge.controller.InputController;
+import bridge.domain.Bridge;
 import bridge.domain.BridgeMaker;
 import bridge.domain.BridgeRandomNumberGenerator;
-import bridge.repository.BridgeRepository;
 import bridge.repository.UserBridgeRepository;
 import bridge.service.GameService;
 import bridge.service.InputValidService;
@@ -19,7 +19,7 @@ public class AppConfig {
 		InputController inputController = new InputController(inputView, new InputValidService(), outputView);
 		UserBridgeRepository userBridgeRepository = new UserBridgeRepository();
 		BridgeMaker bridgeMaker = new BridgeMaker(new BridgeRandomNumberGenerator());
-		GameService gameService = new GameService(bridgeMaker, new BridgeRepository(), userBridgeRepository);
+		GameService gameService = new GameService(bridgeMaker, new Bridge(), userBridgeRepository);
 		return new BridgeGame(inputController, gameService, outputView);
 	}
 }
