+++ src/main/java/bridge/service/GameService.java
@@ -2,8 +2,8 @@ package bridge.service;
 
 import java.util.List;
 
+import bridge.domain.Bridge;
 import bridge.domain.BridgeMaker;
-import bridge.repository.BridgeRepository;
 import bridge.repository.UserBridgeRepository;
 import bridge.repository.dto.UserBridgeStatusDto;
 import bridge.util.OutputViewConst;
@@ -11,24 +11,24 @@ import bridge.util.OutputViewConst;
 public class GameService {
 
 	private final BridgeMaker bridgeMaker;
-	private final BridgeRepository bridgeRepository;
+	private final Bridge bridge;
 	private final UserBridgeRepository userBridgeRepository;
 
-	public GameService(BridgeMaker bridgeMaker, BridgeRepository bridgeRepository,
+	public GameService(BridgeMaker bridgeMaker, Bridge bridge,
 		UserBridgeRepository userBridgeRepository) {
 		this.bridgeMaker = bridgeMaker;
-		this.bridgeRepository = bridgeRepository;
+		this.bridge = bridge;
 		this.userBridgeRepository = userBridgeRepository;
 	}
 
 	public List<String> makeBridge(Integer bridgeSize) {
 		List<String> bridge = bridgeMaker.makeBridge(bridgeSize);
-		bridgeRepository.initBridge(bridge);
+		this.bridge.initBridge(bridge);
 		return bridge;
 	}
 
 	public boolean checkValidSpace(String userSpace, Integer currentSpace) {
-		return bridgeRepository.checkValidSpace(userSpace, currentSpace);
+		return bridge.checkValidSpace(userSpace, currentSpace);
 	}
 
 	public UserBridgeStatusDto saveUserCorrectSpace(String userLocation) {
